+++ src/core/types/MagicalOptions.d.ts
@@ -1,0 +1,7 @@
+type MagicalOptions = {
+  withCondition?: (lead: Record<string, any>) => boolean;
+  domain: string;
+  router: string;
+  debug: boolean;
+};
+export default MagicalOptions;
