--- src/hubspot/common/types.d.ts
@@ -1,7 +1,0 @@
-export interface HubspotMagicalOptions {
-  domain: string;
-  router: string;
-  requiredField: string;
-  formId?: string | string[];
-  debug: boolean;
-}
