+++ src/hubspot/core/types/HubSpotMagicalOptions.d.ts
@@ -1,0 +1,6 @@
+import MagicalOptions from "../../../core/types/MagicalOptions";
+
+export interface HubspotMagicalOptions extends MagicalOptions {
+  formId?: string | string[];
+  requiredField: string;
+}
