--- src/pardot/common/types.d.ts
@@ -1,6 +1,0 @@
-export interface PardotMagicalOptions {
-  maxRetries: number;
-  domainKey: string;
-  routerKey: string;
-  debug: boolean;
-}
