+++ src/pardot/core/helpers.ts
