+++ src/pardot/core/types/PardotMagicalOptions.d.ts
@@ -1,0 +1,7 @@
+export default interface PardotMagicalOptions {
+  withCondition?: (lead: Record<string, any>) => boolean;
+  maxRetries: number;
+  domainKey: string;
+  routerKey: string;
+  debug: boolean;
+}
